requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
ffmpeg-python==0.2.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
kubernetes==24.2.0
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycountry==22.3.5
pycryptodome==3.14.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
slim==0.1
streamlink==4.1.0
tomli==2.0.1
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.2
zipp==3.8.0